# docs/api.yml
openapi: 3.0.0
info:
  title: server programming project API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/user:
    post:
      summary: ユーザー登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                name:
                  type: string
                user_id:
                  type: string
                birthday:
                  type: string
                show_adult_contents:
                  type: boolean
                social_accounts:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      identifier:
                        type: string
      responses:
        "201":
          description: 登録できた場合
        "400":
          description: パラメーターが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: ユーザーIDが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "500":
          description: SQLエラー等の全般的なエラーの発生
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
    delete:
      summary: アカウント削除
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer トークンを入れる（例: Bearer <JWT>）"
      responses:
        "204":
          description: 削除できた場合
        "400":
          description: パラメーターが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: ユーザーIDが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "500":
          description: SQLエラー等の全般的なエラーの発生
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
    patch:
      summary: アカウント編集, ユーザー設定
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer トークンを入れる（例: Bearer <JWT>）"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                name:
                  type: string
                icon:
                  type: string
                  format: binary
                header:
                  type: string
                  format: binary
                social_accounts:
                  type: array
                  items:
                    type: object
                    properties:
                      platform_id:
                        type: integer
                      identifier:
                        type: string
                introduction:
                  type: string
                birthday:
                  type: string
                show_adult_contents:
                  type: boolean
      responses:
        "200":
          description: 編集登録できた場合
        "400":
          description: パラメーターが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: ユーザーIDが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "500":
          description: SQLエラー等の全般的なエラーの発生
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
  /api/user/{userId}:
    get:
      summary: アカウント詳細情報取得
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer トークンを入れる（例: Bearer <JWT>）"
      responses:
        "200":
          description: 認証できた場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  icon_path:
                    type: string
                  header_path:
                    type: string
                  social_accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        platform_id:
                          type: integer
                        identifier:
                          type: string
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        post_id:
                          type: string
                        icon_path:
                          type: string
                        like_count:
                          type: integer
                        content:
                          type: object
                          properties:
                            description:
                              type: string
                            path:
                              type: string
                            alt:
                              type: string
                  liked_posts:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: string
                        name:
                          type: string
                        post_id:
                          type: string
                        icon_path:
                          type: string
                        content:
                          type: object
                          properties:
                            description:
                              type: string
                            path:
                              type: string
                            alt:
                              type: string
        "400":
          description: パラメーターが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: ユーザーIDが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "500":
          description: SQLエラー等の全般的なエラーの発生
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
...
  /api/like/{postId}:
    post:
      summary: いいね登録
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer トークンを入れる（例: Bearer <JWT>）"
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: 登録できた場合
        "400":
          description: パラメーターが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: ユーザーIDが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "500":
          description: SQLエラー等の全般的なエラーの発生
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string

    delete:
      summary: いいね削除
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer トークンを入れる（例: Bearer <JWT>）"
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 削除できた場合
        "400":
          description: パラメーターが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: ユーザーIDが不正の場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "500":
          description: SQLエラー等の全般的なエラーの発生
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
...